move visualization code to separate file
